board/ci20: Implement creation of a basic sdcard image for ci20
Add a genimage.cfg and post-image.sh to create a working sdcard.img for
the ci20 hardware. We also need a uboot-env.txt to create the partition
for the uboot environment.

Update the board/ci20/readme.txt with the related information. Remove
the tftp netboot description from the readme as it is already to
complicated for a basic bootable example.

Fix the toolchain-wrapper to support more recent GCC versions concerning
the MIPS XBurst architecture used by ci20.

Signed-off-by: Johannes Schmitz <johannes.schmitz1@gmail.com>